+++ src/platform/android/mod.rs
@@ -160,7 +160,7 @@ impl MonitorId {
     }
 
     #[inline]
-    pub fn get_position(&self) -> (u32, u32) {
+    pub fn get_position(&self) -> (i32, i32) {
         // Android assumes single screen
         (0, 0)
     }