+++ src/platform/emscripten/mod.rs
@@ -34,7 +34,7 @@ impl MonitorId {
     }
 
     #[inline]
-    pub fn get_position(&self) -> (u32, u32) {
+    pub fn get_position(&self) -> (i32, i32) {
         unimplemented!()
     }