+++ src/platform/macos/monitor.rs
@@ -50,7 +50,7 @@ impl MonitorId {
     }
 
     #[inline]
-    pub fn get_position(&self) -> (u32, u32) {
+    pub fn get_position(&self) -> (i32, i32) {
         unimplemented!()
     }